pyproject.toml
[tool.poetry]
authors = ["yiskw713 <i.yuchi713@gmail.com>"]
description = "pytorch implementation template"
license = "MIT"
name = "pytorch_videoloader"
version = "0.1.0"

[tool.poetry.dependencies]
Pillow = "^8.0.1"
PyYAML = "^5.3.1"
h5py = "^3.2.1"
matplotlib = "^3.4.2"
numpy = "^1.19.4"
pandas = "^1.1.4"
python = "^3.8"
torch = "^1.7.0"
torchvision = "^0.8.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
mypy = "^0.790"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
